requirements.txt
# SAP CPI Specification Generator - Dependencies
# Python 3.10+

# Core document generation
python-docx>=1.1.0

# AI generation
google-genai>=1.0.0

# Environment configuration
python-dotenv>=1.0.0

# CLI color and terminal UI
rich>=13.7.0
questionary>=2.0.1

# XML processing (for XSD schemas)
lxml>=5.0.0

# BPMN parsing support
bpmn_python>=0.0.18

# Optional - used for imports but not core functionality
# networkx>=3.0
# matplotlib>=3.8.0
# requests>=2.31.0
